more fixes

+++ cirq-core/cirq/transformers/analytical_decompositions/two_qubit_to_fsim.py
@@ -252,9 +252,7 @@ def _fix_single_qubit_gates_around_kak_interaction(
         A list of operations whose kak decomposition approximately equals the
         desired kak decomposition.
     """
-    actual = linalg.kak_decomposition(
-        circuits.Circuit(operations).unitary(qubit_order=qubits, dtype=np.complex128)
-    )
+    actual = linalg.kak_decomposition(circuits.Circuit(operations).unitary(qubit_order=qubits))
 
     def dag(a: np.ndarray) -> np.ndarray:
         return np.transpose(np.conjugate(a))
